Update EVT, boards.txt, platform.txt

+++ copy/libraries/noneOS/examples/common/c_main/c_main.c
@@ -1,13 +1,13 @@
 #include "debug.h"
 
 void c_main() {
-  u8 i = 0;
+  uint8_t i = 0;
 
   SystemCoreClockUpdate();
   Delay_Init();
   USART_Printf_Init(115200);
-  printf("SystemClk:%d\r\n", SystemCoreClock);
-  printf("ChipID:%08x\r\n", DBGMCU_GetCHIPID());
+  printf("SystemClk:%u\r\n", (unsigned int)SystemCoreClock);
+  printf("ChipID:%08x\r\n", (unsigned int)DBGMCU_GetCHIPID());
   printf("C main TEST\r\n");
 
   while (1) {
